Fixes Kane.Topic.create/1 returning a GCP error
When calling Kane.Topic.create/1 I get the following GCP error:

```
{
  "error": {
    "code": 400,
    "message": "Invalid JSON payload received. Unknown name \\\"\\\": Root element must be a message.",
    "status": "INVALID_ARGUMENT",
    "details": [
      {
        "@type": "type.googleapis.com/google.rpc.BadRequest",
        "fieldViolations": [
          {
            "description": "Invalid JSON payload received. Unknown name \\\"\\\": Root element must be a message."
          }
        ]
      }
    ]
  }
}
```

This was due to the payload on the HTTP request body not being empty, but instead this string: `\"\"`, which was generated by the Poison.encode!/1 function being used with an empty string, i.e.:

```
iex(1)> Poison.encode!("")
"\"\""
```

My solution for this issue is to only use Poison.encode!/1 when the data being passed in is not an empty string. After doing this Kane.Topic.create/1 works as expected.