Make type spec for Kane.Message struct more permissive
The type spec for the Kane.Message struct required `id`, `ack_id`, and
`publish_time` to be of type String.t. However, the default value for
these fields is `nil`, which is not of type String.t. This caused
Dialyzer errors whenever creating new messages that used the default
values for these fields. By changing the type specs for these fields to
allow for `nil`, these Dialyzer errors are resolved.