Add expiration and message retention options to Subscription
Hello!

In Cloud Pubsub we can specify for how long a message will be retained (https://cloud.google.com/pubsub/docs/reference/rest/v1/projects.subscriptions/create#request-body)

We're using these changes on our production system and has been working pretty well.

Our use case:

We create many dynamic subscriptions during the day.
Old subscriptions aren't used, so it ended up with some leftover long-lived subscriptions that we don't use.

It can be dangerous because we have a limit on how many subscriptions we
can have. The default expiration of an inactive subscription is 31 days, we want to reduce it to t
the minimum value that is 1 day.

However, according to the docs, the subscription expiration should
always be greater than the message retention duration. The default
message retention is 7 days, so we need to reduce it together.